feat(terminal/ios): separate mappers
